feat: refactored book player into subfolder and changed style of progress bar
