+++ components/BookPlayer/BookTextComponent.tsx
