+++ pages/demo.tsx
@@ -1,4 +1,4 @@
-import BookPlayer from "../components/BookPlayer";
+import BookPlayer from "../components/BookPlayer/BookPlayer";
 
 export default function Demo() {
     return (
